Replace File with generic read/write + seek traits in Compression
The main reason for this change is to be able to use std::io::Cursor, an
in-memory buffer which can be used as a replacement for File. This is
useful because we want to compress blobs in-memory and only write them
to disk if they take up less space than the uncompressed blob.

Signed-off-by: Ariel Miculas <amiculas@cisco.com>